Tambah dikit, gatau apa aja lupa

+++ controllers/invoiceController.php
@@ -129,7 +129,9 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
       $pathDb = $crud->showData("SELECT bukti_pengiriman FROM transaksi WHERE kode_pesanan='" . $_POST['id'] . "'");
       $pathImg = "../images/bukti_pengiriman/" . $pathDb[0]['bukti_pengiriman'];
       $crud->execute("UPDATE transaksi SET status_pengiriman='kirim', bukti_pengiriman=NULL WHERE kode_pesanan = '" . $_POST['id'] . "'");
-      unlink($pathImg);
+      if (file_exists($pathImg)) {
+        unlink($pathImg);
+      }
 
       $response = array(
         'status' => 'success',
